Turns out that the server does not use the md5 of files when performing multi-part uploads. And the creation of the unused md5 digest is [very expensive](https://github.com/modal-labs/modal-client/pull/2649#issuecomment-2541013862).

This PR omits sending a real md5 for files larger than 1GiB, which we know** will always be multipart uploaded.

** We'll likely remove md5 entirely and only use the sha256 hash. Until then we'll make sure not to raise the multipart upload threshold higher. If anything we're [lowering it](https://github.com/modal-labs/modal/pull/18208). 

TODO:
- [x] Test volume put 2GiB file
- [x] Test volume put 16MiB file
- [x] Test mount 2GiB file
- [x] Test mount 16MiB file

## Benchmark

TL;DR: We save ~2s per GiB by not doing the md5 hash. For a 64GiB file that's ~2 minutes before we even start uploading. For a 256GiB file that's ~8 minutes. 

The below benchmark uploads a 16GiB file.

```
MODAL_LOGLEVEL=DEBUG modal volume put dano-test data-16g  --force >/dev/null
```

### Before
```
[ThreadPoolExecutor-0_0] 2024-12-13T15:30:34+0100 get_sha256_hex took 9.462s
[Thread-1 (thread_inner)] 2024-12-13T15:31:01+0100 get_upload_hashes took 27.037s (get_md5_base64)
```

### After
```
[ThreadPoolExecutor-0_0] 2024-12-13T15:29:25+0100 get_upload_hashes took 9.033s (dict_keys(['sha256']))
[Thread-1 (thread_inner)] 2024-12-13T15:29:25+0100 get_upload_hashes took 0.000s (dict_keys([]))
```



---

Check these boxes or delete any item (or this section) if not relevant for this PR.

- [ ] Client+Server: this change is compatible with old servers
- [ ] Client forward compatibility: this change ensures client can accept data intended for later versions of itself

Note on protobuf: protobuf message changes in one place may have impact to
multiple entities (client, server, worker, database). See points above.

---

</details>

## Changelog

<!--
If relevant, include a brief user-facing description of what's new in this version.

Format the changelog updates using bullet points.
See https://modal.com/docs/reference/changelog for examples and try to use a consistent style.

Provide short code examples, indented under the relevant bullet point, if they would be helpful.
Cross-linking to relevant documentation is also encouraged.
-->